// src/api/models/OrganizationMemberModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OrganizationMemberModel = {
  organizationId?: string;
  email: string;
  role: string;
  password?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  dateOfBirth?: string | null;
  profilePictureUrl?: string | null;
  bio?: string | null;
};